Refactor and enhance GPTQ functionality
- Updated import paths for GPTQ-related modules in `qwen2moe.py`.
- Added a new test script `load_demo.py` to simulate and load GPTQ weights from a safetensors file.
- Introduced `matmul_test.py` to verify the correctness of the GPTQ matrix multiplication implementation.
- Created a mock class `MockGPTQLinear` for testing purposes in `mock_GPTQLinear.py`.
- Implemented `verify_fused_gate_up.py` to validate the fused gate and up operations.
- Added `verify_gptq_matmul.py` to ensure the correctness of GPTQ matrix multiplication.
- Developed `gptq_quantize_kernel.py` to handle quantization operations for GPTQ.
- Implemented `matmul_gptq.py` for optimized matrix multiplication using GPTQ.
- Created `matmul_kernel.py` for general matrix multiplication operations.
- Added `softmax_impl.py` to implement a custom softmax function using Triton.
- Updated `run_profile.py` to reflect new import paths for the matmul_gptq function.

+++ nanovllm/tests/verify_fused_gate_up.py
@@ -3,8 +3,8 @@ import triton
 import triton.language as tl
 
 from nanovllm.layers.activation import SiluAndMul
-from test.mock.mock_GPTQLinear import MockGPTQLinear
-from triton_impl.gptq_quantize_kernel import fused_gate_up
+from nanovllm.tests.mock.mock_GPTQLinear import MockGPTQLinear
+from nanovllm.triton_impl.gptq_quantize_kernel import fused_gate_up
 
 
 def prepare_data(device="cuda"):
@@ -145,8 +145,8 @@ def verify_gate_up(atol=1e-2, rtol=1e-1):
     g_w, g_z, g_s = prepare_data()
     u_w, u_z, u_s = prepare_data()
 
-    gate = unpack_wzs_torch(g_w, g_z, g_s)
-    up = unpack_wzs_torch(u_w, u_z, u_s)
+    gate = unpack_wzs_torch(g_w, g_z, g_s).to(torch.float16)
+    up = unpack_wzs_torch(u_w, u_z, u_s).to(torch.float16)
 
     gate_up = torch.cat(
         (hidden_size @ gate, hidden_size @ up),
